fix(server): unable to do first setup

+++ apps/server/src/routes/session_parser.ts
@@ -1,4 +1,5 @@
 import sql from "../services/sql.js";
+import sqlInit from "../services/sql_init.js";
 import session, { Store } from "express-session";
 import sessionSecret from "../services/session_secret.js";
 import config from "../services/config.js";
@@ -20,6 +21,10 @@ export const SESSION_COOKIE_EXPIRY = 24 * 60 * 60 * 1000; // 24 hours
 export class SQLiteSessionStore extends Store {
 
     get(sid: string, callback: (err: any, session?: session.SessionData | null) => void): void {
+        if (!sqlInit.isDbInitialized()) {
+            return callback(null, null);
+        }
+
         try {
             const data = sql.getValue<string>(/*sql*/`SELECT data FROM sessions WHERE id = ?`, sid);
             let session = null;
@@ -34,6 +39,10 @@ export class SQLiteSessionStore extends Store {
     }
 
     set(id: string, session: session.SessionData, callback?: (err?: any) => void): void {
+        if (!sqlInit.isDbInitialized()) {
+            return callback?.();
+        }
+
         try {
             const expires = session.cookie?.expires
                 ? new Date(session.cookie.expires).getTime()
@@ -53,6 +62,10 @@ export class SQLiteSessionStore extends Store {
     }
 
     destroy(sid: string, callback?: (err?: any) => void): void {
+        if (!sqlInit.isDbInitialized()) {
+            return callback?.();
+        }
+
         try {
             sql.execute(/*sql*/`DELETE FROM sessions WHERE id = ?`, sid);
             callback?.();
@@ -63,6 +76,10 @@ export class SQLiteSessionStore extends Store {
     }
 
     touch(sid: string, session: session.SessionData, callback?: (err?: any) => void): void {
+        if (!sqlInit.isDbInitialized()) {
+            return callback?.();
+        }
+
         // For now it's only for session cookies ("Remember me" unchecked).
         if (session.cookie?.expires) {
             callback?.();
